Rename to themedComponents, add FullWidthSpace
