+++ src/containers/notFound/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { Link } from 'react-router-dom';
 import { Routes } from '../../App';
-import { ContentContainer } from '../../components/styledComponents';
+import { ContentContainer } from '../../components/themedComponents';
 import { Typography } from 'antd';
 const { Title } = Typography;
